+++ src/test/scala/com/sagebear/bigrussianboss/bot/ObedientBot.scala
@@ -6,15 +6,14 @@ import com.sagebear.bigrussianboss.bot.SensorsAndActuators.CanNotDoThis
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.Try
 
-class ObedientBot(private var phrases: List[String]) extends SensorsAndActuators {
+class ObedientBot(private val phrases: List[String]) extends SensorsAndActuators {
   override def observe(text: String)(a: Script.Action)(implicit ec: ExecutionContext): Future[SensorsAndActuators] = {
-    Future(new ObedientBot(phrases))
+    Future(new ObedientBot(phrases.tail))
   }
 
   override def act(a: Script.Action)(implicit ec: ExecutionContext): Future[String] = {
     this.phrases match {
-      case head :: tail =>
-        this.phrases = tail
+      case head :: _ =>
         Future(head)
       case _ =>
         Future.failed(CanNotDoThis)
@@ -24,6 +23,5 @@ class ObedientBot(private var phrases: List[String]) extends SensorsAndActuators
 
 object ObedientBot {
   def client(phrases: String*): Try[ObedientBot] = Try(new ObedientBot(phrases.toList))
-
   def operator(phrases: String*): Try[ObedientBot] = Try(new ObedientBot(phrases.toList))
 }
